[update] create the directory elfload and split the file into it
